// src/api/models/chatQuery.ts
/**
 * Generated by orval v6.15.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AppUser } from "./appUser";

export interface ChatQuery {
  id: number;
  timestamp: string | Date;
  user: AppUser;
  queryContent: string;
  sanitizedQueryContent: string;
}